[Fix] 데이터 있을시 음식 검색창 스킵 작업 진행중
