+++ src/main/ayofront/components/record/MealCard1.js
@@ -45,17 +45,17 @@ const MealCard1 = ({
           <View>
             <Text style={styles.mealTime}> </Text>
             <Text style={styles.nutrientValue}>
-              {carb === null || carb === undefined ? "-" : Math.floor(carb)} g
+              {carb === null || carb === undefined ? 0 : Math.floor(carb)} g
             </Text>
             <Text style={styles.nutrientValue}>
-              {protein === null || protein === undefined ? "-" : Math.floor(protein)} g
+              {protein === null || protein === undefined ? 0 : Math.floor(protein)} g
             </Text>
             <Text style={styles.nutrientValue}>
-              {fat === null || fat === undefined ? "-" : Math.floor(fat)} g
+              {fat === null || fat === undefined ? 0 : Math.floor(fat)} g
             </Text>
             <Text style={styles.carloriesValue}>
               {carb === null || protein === null || fat === null || carb === undefined || protein === undefined || fat === undefined
-                ? "- kcal"
+                ? "0 kcal"
                 : Math.floor(carb * 4 + protein * 4 + fat * 9) + " kcal"}
             </Text>
           </View>
