+++ src/main/ayofront/screens/record/RecordScreen.js
@@ -18,7 +18,7 @@ const { debuggerHost } = Constants.manifest2.extra.expoGo;
 const uri = `http://${debuggerHost.split(":").shift()}:8080`;
 
 function RecordScreen({ navigation }) {
-  const { updateMealType, formattedYYMMDD} = useMealContext();
+  const { mealType, addItemToMealList, updateMealType, formattedYYMMDD, cleanMealList} = useMealContext();
 
   // 카드를 클릭할때 mealType 받아서 mealContext에 저장
   const setMealType = (mealType) => {
@@ -87,15 +87,37 @@ function RecordScreen({ navigation }) {
     getImg();
   }, []);
 
+  //SearchModal을 거치지 않고 mealMain페이지로 진입시 Sever에서 해당 날자 식단을 가져와서 MealContext에 저장
+  const getMealByTypeAndDate = () => {
+    axios
+      .get(`${uri}/api/meal/type`,
+        {
+          params: {
+            mealType: mealType,
+            date: formattedYYMMDD,
+          }
+        })
+      .then((response) => {
+        console.log('GET서버 => RecordMain.js:', response.data);
+        addItemToMealList(response.data);
+      })
+      .catch(() => {
+        console.log("getMealDataByTypeAndDate error..");
+      });
+  };
+
   const handleCardPress = (cardData) => {
     setMealType(cardData.mealType.toLowerCase());
     setPhotoUri(img.find((item) => item.fType === cardData.mealType.toLowerCase())?.fImg);
     setPhotoId(img.find((item) => item.fType === cardData.mealType.toLowerCase())?.fNo);
     
     if (cardData.meals) {
+      cleanMealList([]);
+      getMealByTypeAndDate();
       navigation.push("RecordMain"); // 음식 데이터가 있으면 RecordMain 화면으로 이동
     } else {
-      openSearchModal(); // 없으면 openSearchModal 실행
+      cleanMealList();
+      openSearchModal(); // 음식 데이터가 없으면 openSearchModal 실행
     }
   };
